[master]活动保存支持活动渠道信息保存

+++ nas-common/src/main/java/com/minyan/nascommon/po/ActivityChannelTempPO.java
@@ -22,7 +22,7 @@ public class ActivityChannelTempPO implements Serializable {
     /**
      * 活动id
      */
-    private Long activityId;
+    private Integer activityId;
 
     /**
      * 渠道名称
